# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TeamZeit API
  version: 0.2.0
  description: Tenant-safe direct work-interval time tracking.
servers: [{ url: /api/v1 }]
security: [{ bearerAuth: [] }]
paths:
  /month-closures/{membershipId}/{month}:
    get:
      operationId: getMonthClosure
      tags: [MonthClosing]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/MembershipId' }, { name: month, in: path, required: true, schema: { type: string, pattern: '^\d{4}-(0[1-9]|1[0-2])$' } }]
      responses: { '200': { description: Current open or closed status }, '403': { $ref: '#/components/responses/Forbidden' } }
  /month-closures/close:
    post:
      operationId: closeMonth
      tags: [MonthClosing]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MonthClosureCommand' } } } }
      responses: { '200': { description: Month closed and audited }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' } }
  /month-closures/reopen:
    post:
      operationId: reopenMonth
      tags: [MonthClosing]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MonthClosureCommand' } } } }
      responses: { '200': { description: Month reopened and audited }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' } }
  /employees:
    get:
      operationId: listEmployees
      tags: [Identity]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }]
      responses: { '200': { description: Tenant employee administration list }, '403': { $ref: '#/components/responses/Forbidden' } }
    post:
      operationId: createEmployee
      tags: [Identity]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateEmployee' } } } }
      responses: { '201': { description: Invited employee record created without sending an invitation }, '403': { $ref: '#/components/responses/Forbidden' } }
  /employees/{membershipId}/invitation:
    post:
      operationId: sendEmployeeInvitation
      tags: [Identity]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/MembershipId' }]
      responses: { '200': { description: Invitation sent or resent }, '403': { $ref: '#/components/responses/Forbidden' } }
  /employees/invitations:
    post:
      operationId: inviteEmployee
      tags: [Identity]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/InviteEmployee' } } } }
      responses: { '201': { description: Invitation and invited membership created }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' } }
  /employees/{membershipId}/deactivate:
    post:
      operationId: deactivateEmployee
      tags: [Identity]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/MembershipId' }]
      responses: { '200': { description: Membership deactivated and audited }, '403': { $ref: '#/components/responses/Forbidden' } }
  /employees/{membershipId}/assignment:
    patch:
      operationId: updateEmployeeAssignment
      tags: [Identity]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/MembershipId' }]
      responses: { '200': { description: Role, team, and work policy assignment updated and audited }, '403': { $ref: '#/components/responses/Forbidden' } }
  /attendance/today:
    get:
      operationId: getTodayAttendance
      tags: [TimeTracking]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }]
      responses: { '200': { description: Current day including intervals and derived totals, content: { application/json: { schema: { $ref: '#/components/schemas/TodayAttendance' } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
  /attendance/days/{workDate}:
    get:
      operationId: getDailyAttendance
      tags: [TimeTracking]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { name: workDate, in: path, required: true, schema: { type: string, format: date } }]
      responses: { '200': { description: Daily interval overview, content: { application/json: { schema: { $ref: '#/components/schemas/DailyAttendance' } } } } }
  /attendance/months/{month}:
    get:
      operationId: getMonthlyAttendance
      tags: [TimeTracking]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { name: month, in: path, required: true, schema: { type: string, pattern: '^\d{4}-\d{2}$' } }]
      responses: { '200': { description: Monthly totals, content: { application/json: { schema: { $ref: '#/components/schemas/MonthlyAttendance' } } } } }
  /attendance/employees/{membershipId}/work-rule:
    put:
      operationId: setEmployeeWorkRule
      tags: [TimeTracking]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/MembershipId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SetEmployeeWorkRule' } } } }
      responses: { '200': { description: Effective-dated employee work rule created and audited }, '403': { $ref: '#/components/responses/Forbidden' } }
  /attendance/commands/clock-in:
    post:
      operationId: clockIn
      tags: [TimeTracking]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { $ref: '#/components/responses/ClockCommand' }, '409': { $ref: '#/components/responses/Conflict' } }
  /attendance/commands/clock-out:
    post:
      operationId: clockOut
      tags: [TimeTracking]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { $ref: '#/components/responses/ClockCommand' }, '409': { $ref: '#/components/responses/Conflict' } }
  /attendance/commands/break-start:
    post:
      deprecated: true
      operationId: deprecatedBreakStartAsClockOut
      tags: [TimeTracking]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { $ref: '#/components/responses/ClockCommand' } }
  /attendance/commands/break-end:
    post:
      deprecated: true
      operationId: deprecatedBreakEndAsClockIn
      tags: [TimeTracking]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { $ref: '#/components/responses/ClockCommand' } }
  /attendance/sessions:
    get:
      operationId: listOwnIntervals
      tags: [TimeTracking]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { name: from, in: query, required: true, schema: { type: string, format: date } }, { name: to, in: query, required: true, schema: { type: string, format: date } }]
      responses: { '200': { description: Own intervals } }
    post:
      operationId: createOwnInterval
      tags: [TimeTracking]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateInterval' } } } }
      responses: { '201': { description: Interval created immediately, content: { application/json: { schema: { $ref: '#/components/schemas/WorkSession' } } } }, '409': { $ref: '#/components/responses/Conflict' } }
  /attendance/sessions/{sessionId}:
    put:
      operationId: updateOwnInterval
      tags: [TimeTracking]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/SessionId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateInterval' } } } }
      responses: { '200': { description: Interval updated immediately }, '409': { $ref: '#/components/responses/Conflict' } }
    delete:
      operationId: archiveOwnInterval
      tags: [TimeTracking]
      parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/SessionId' }, { name: expectedVersion, in: query, required: true, schema: { type: integer, minimum: 1 } }]
      responses: { '200': { description: Interval archived immediately }, '409': { $ref: '#/components/responses/Conflict' } }
components:
  securitySchemes: { bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT } }
  parameters:
    OrganizationId: { name: X-Organization-Id, in: header, required: true, schema: { type: string, format: uuid }, description: Tenant selector verified against the authenticated membership. }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, format: uuid } }
    SessionId: { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
    MembershipId: { name: membershipId, in: path, required: true, schema: { type: string, format: uuid } }
  responses:
    ClockCommand: { description: Server-time clock result, content: { application/json: { schema: { type: object, required: [serverTime, session], properties: { serverTime: { type: string, format: date-time }, session: { $ref: '#/components/schemas/WorkSession' } } } } } }
    Forbidden: { description: Active tenant membership or ownership missing }
    Conflict: { description: Invalid state, stale version, overlap, duplicate action, or closed period }
  schemas:
    MonthClosureCommand:
      type: object
      additionalProperties: false
      required: [membershipId, month, reason]
      properties: { membershipId: { type: string, format: uuid }, month: { type: string, pattern: '^\d{4}-(0[1-9]|1[0-2])$' }, reason: { type: string, minLength: 3, maxLength: 500 } }
    InviteEmployee:
      type: object
      additionalProperties: false
      required: [email, role]
      properties: { email: { type: string, format: email }, role: { type: string, enum: [admin, manager, employee, auditor] }, teamId: { type: string, format: uuid }, workPolicyId: { type: string, format: uuid } }
    CreateEmployee:
      type: object
      additionalProperties: false
      required: [email, role]
      properties: { email: { type: string, format: email }, role: { type: string, enum: [admin, manager, employee, auditor] } }
    WorkSession:
      type: object
      required: [id, organizationId, membershipId, workDate, startedAt, state, source, version]
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        membershipId: { type: string, format: uuid }
        workDate: { type: string, format: date }
        startedAt: { type: string, format: date-time }
        endedAt: { type: string, format: date-time }
        workedMinutes: { type: integer, minimum: 0 }
        calculationEndedAt: { type: string, format: date-time }
        issue: { type: string, enum: [missing_clock_out] }
        state: { type: string, enum: [not_started, working, completed, requires_action] }
        source: { type: string, enum: [clock, manual, admin_import] }
        version: { type: integer, minimum: 1 }
    CreateInterval:
      type: object
      additionalProperties: false
      required: [workDate, startedAt, endedAt]
      properties: { workDate: { type: string, format: date }, startedAt: { type: string, format: date-time }, endedAt: { type: string, format: date-time } }
    UpdateInterval:
      allOf: [{ $ref: '#/components/schemas/CreateInterval' }, { type: object, required: [expectedVersion], properties: { expectedVersion: { type: integer, minimum: 1 } } }]
    SetEmployeeWorkRule:
      type: object
      additionalProperties: false
      required: [effectiveFrom, weekdayMinutes]
      properties:
        effectiveFrom: { type: string, format: date }
        weekdayMinutes: { type: object, required: [monday, tuesday, wednesday, thursday, friday, saturday, sunday], additionalProperties: false, properties: { monday: { type: integer, minimum: 0, maximum: 1440 }, tuesday: { type: integer, minimum: 0, maximum: 1440 }, wednesday: { type: integer, minimum: 0, maximum: 1440 }, thursday: { type: integer, minimum: 0, maximum: 1440 }, friday: { type: integer, minimum: 0, maximum: 1440 }, saturday: { type: integer, minimum: 0, maximum: 1440 }, sunday: { type: integer, minimum: 0, maximum: 1440 } } }
        breakThresholdMinutes: { type: integer, default: 360 }
        minimumBreakMinutes: { type: integer, default: 30 }
    DailyAttendance:
      type: object
      required: [workDate, state, sessions, workedMinutes, breakMinutes, recordedBreakMinutes, automaticBreakMinutes, plannedMinutes, balanceMinutes, isHoliday, requiresAction]
      properties: { workDate: { type: string, format: date }, state: { type: string, enum: [not_started, working, completed, requires_action] }, sessions: { type: array, items: { $ref: '#/components/schemas/WorkSession' } }, workedMinutes: { type: integer }, breakMinutes: { type: integer }, recordedBreakMinutes: { type: integer }, automaticBreakMinutes: { type: integer }, plannedMinutes: { type: integer }, balanceMinutes: { type: integer }, isHoliday: { type: boolean }, requiresAction: { type: boolean } }
    TodayAttendance:
      allOf: [{ $ref: '#/components/schemas/DailyAttendance' }, { type: object, required: [serverTime], properties: { serverTime: { type: string, format: date-time }, activeSession: { $ref: '#/components/schemas/WorkSession' } } }]
    MonthlyAttendance:
      type: object
      required: [month, days, workedMinutes, breakMinutes, plannedMinutes, balanceMinutes, requiresAction]
      properties: { month: { type: string }, days: { type: array, items: { $ref: '#/components/schemas/DailyAttendance' } }, workedMinutes: { type: integer }, breakMinutes: { type: integer }, plannedMinutes: { type: integer }, balanceMinutes: { type: integer }, requiresAction: { type: boolean } }