# services/api/src/month-closing/routes.ts
import type { ChangeMonthClosureRequest, MembershipRole, UUID } from "@teamzeit/contracts";
import type { FastifyInstance, FastifyReply, FastifyRequest } from "fastify";
import { z } from "zod";
import type { ApiConfig } from "../config/env.js";
import { IdentityError, resolveCurrentContext, type IdentityContextDependencies } from "../identity/context.js";
import { TimeTrackingError } from "../time-tracking/errors.js";
import { MonthClosingService } from "./service.js";
import type { MonthClosingActor } from "./types.js";
const uuid = z.string().uuid(); const month = z.string().regex(/^\d{4}-(0[1-9]|1[0-2])$/u); const body = z.object({ membershipId: uuid, month, reason: z.string().trim().min(3).max(500) }) satisfies z.ZodType<ChangeMonthClosureRequest>;
export interface MonthClosingRouteDependencies { service: MonthClosingService; identity?: IdentityContextDependencies; } interface Context { actor: MonthClosingActor; role: MembershipRole; }
export function registerMonthClosingRoutes(app: FastifyInstance, config: ApiConfig, dependencies: MonthClosingRouteDependencies): void {
  app.get("/api/v1/month-closures/:membershipId/:month", (request, reply) => handle(request, reply, config, dependencies.identity, async (context) => { const parsed = parse(z.object({ membershipId: uuid, month }), request.params); if (!isAdmin(context.role) && parsed.membershipId !== context.actor.membershipId) throw new TimeTrackingError("FORBIDDEN", "Der Monatsabschluss eines anderen Mitarbeiters ist nicht sichtbar."); return dependencies.service.getStatus(context.actor, parsed.membershipId, parsed.month); }));
  app.post("/api/v1/month-closures/close", (request, reply) => command(request, reply, config, dependencies, "close")); app.post("/api/v1/month-closures/reopen", (request, reply) => command(request, reply, config, dependencies, "reopen"));
}
function command(request: FastifyRequest, reply: FastifyReply, config: ApiConfig, dependencies: MonthClosingRouteDependencies, operation: "close" | "reopen") { return handle(request, reply, config, dependencies.identity, (context) => { if (!isAdmin(context.role)) throw new TimeTrackingError("FORBIDDEN", "Nur Administratoren dürfen Monate schließen oder öffnen."); const value = parse(body, request.body); return dependencies.service[operation](context.actor, key(request), value.membershipId, value.month, value.reason); }); }
async function handle<T>(request: FastifyRequest, reply: FastifyReply, config: ApiConfig, identity: IdentityContextDependencies | undefined, action: (context: Context) => Promise<T>): Promise<T | FastifyReply> { try { return await action(await context(config, request, identity)); } catch (error) { if (error instanceof IdentityError) return reply.status(error.statusCode).send({ error: { code: error.code, message: error.message, requestId: request.id } }); if (error instanceof TimeTrackingError) return reply.status(error.code === "FORBIDDEN" ? 403 : error.code === "NOT_FOUND" ? 404 : error.code === "VALIDATION_ERROR" ? 400 : error.code === "INTERNAL_ERROR" ? 500 : 409).send({ error: { code: error.code, message: error.message, ...(error.field ? { field: error.field } : {}), requestId: request.id } }); request.log.error(error); return reply.status(500).send({ error: { code: "INTERNAL_ERROR", message: "Interner Fehler.", requestId: request.id } }); } }
async function context(config: ApiConfig, request: FastifyRequest, identity?: IdentityContextDependencies): Promise<Context> { const organizationId = one(request.headers["x-organization-id"]); if (!organizationId || !uuid.safeParse(organizationId).success) throw new TimeTrackingError("FORBIDDEN", "Organisation erforderlich."); const current = await resolveCurrentContext(config, request.headers.authorization, identity); const membership = current.memberships.find((item) => item.organization.id === organizationId && item.status === "active"); if (!membership) throw new TimeTrackingError("FORBIDDEN", "Keine aktive Mitgliedschaft für diese Organisation."); return { role: membership.role, actor: { organizationId, membershipId: membership.id, userId: current.user.id } }; }
function key(request: FastifyRequest): UUID { const value = one(request.headers["idempotency-key"]); if (!value || !uuid.safeParse(value).success) throw new TimeTrackingError("VALIDATION_ERROR", "Gültiger Idempotency-Key erforderlich.", "Idempotency-Key"); return value; } function parse<T>(schema: z.ZodType<T>, value: unknown): T { const result = schema.safeParse(value); if (!result.success) throw new TimeTrackingError("VALIDATION_ERROR", "Anfragedaten sind ungültig.", result.error.issues[0]?.path.join(".")); return result.data; } function isAdmin(role: MembershipRole) { return role === "admin" || role === "owner"; } function one(value: string | string[] | undefined) { return Array.isArray(value) ? value[0] : value; }